Arijit1.5

+++ src/app/signup-bidder/signup-bidder.component.ts
@@ -1,8 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { User } from '../models/user.model';
-import { FormGroup, FormControl , Validators } from '@angular/forms';
+//import { FormGroup, FormControl , Validators } from '@angular/forms';
 import { UserService } from '../services/user.service';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { ConfirmedValidator } from './confirmed.validator';
+
+
 
 @Component({
   selector: 'app-signup-bidder',
@@ -12,50 +16,71 @@ import { UserService } from '../services/user.service';
 export class SignupBidderComponent implements OnInit {
 
   user = new User();
-  form2 : FormGroup;
+  form2: FormGroup;
+  submitted = false;
+  confirmpassword: string;
 
-
-  constructor(private service:UserService,private router: Router) { }
+  constructor(private service: UserService, private router: Router, private formBuilder: FormBuilder) { }
 
   ngOnInit() {
-
-      this.form2=new FormGroup({
-      full_name: new FormControl('',[Validators.required, Validators.pattern('[a-zA-Z][a-zA-Z]+')]),
-      email: new FormControl('',[Validators.required, Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,}$')]),
-      contact_no: new FormControl('',[Validators.required, Validators.pattern('[0-9]{10}')]),
-      addr_1: new FormControl('',[Validators.required]),
-      addr_2: new FormControl('',[Validators.required]),
-      city: new FormControl('',[Validators.required, Validators.pattern('[a-zA-Z][a-zA-Z]+')]),
-      state:new FormControl('',[Validators.required, Validators.pattern('[a-zA-Z][a-zA-Z]+')]),
-      pincode:new FormControl('',[Validators.required, Validators.pattern('[0-9]{6}')]),
-      land_pincode:new FormControl('',[Validators.required, Validators.pattern('[0-9]{6}')]),
-      land_addr:new FormControl('',[Validators.required]),
-      land_area:new FormControl('',[Validators.required, Validators.pattern('[a-zA-Z][a-zA-Z]+')]),
-      account_no:new FormControl('',[Validators.required, Validators.pattern('[0-9]{13}')]),
-      ifsc_no:new FormControl('',[Validators.required, Validators.pattern('[0-9]{6}')]),
-
-
-    })
+    this.form2 = this.formBuilder.group({
+      full_name: ['', [Validators.required, Validators.pattern('[a-zA-Z ][a-zA-Z ]*$')]],
+      email: ['', [Validators.required, Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,}$')]],
+      contact_no: ['', [Validators.required, Validators.pattern('[0-9]{10}')]],
+      addr_1: ['', [Validators.required]],
+      addr_2: ['', [Validators.required]],
+      city: ['', [Validators.required, Validators.pattern('[a-zA-Z][a-zA-Z]+')]],
+      state: ['', [Validators.required, Validators.pattern('[a-zA-Z][a-zA-Z]+')]],
+      pincode: ['', [Validators.required]],
+      account_no: ['', [Validators.required]],
+      ifsc_no: ['', [Validators.required]],
+      password1: ['', [Validators.required]],
+      password2:['',[Validators.required]]
+    // }, {
+    //   validator: ConfirmedValidator('password1', 'password2')
+    }
+    )
   }
 
+  get f() { return this.form2.controls; }
 
-  signupbidder(){
 
-    this.user.status="PENDING";
-    this.user.user_type="BIDDER";
-    localStorage.setItem('usertype',this.user.user_type);
-    
-  console.log(this.user);
 
-    this.service.signupfarmer(this.user).subscribe(data=>{
-        
+  signupbidder() {
+
+    console.log(this.user.password);
+    console.log(this.confirmpassword);
+    this.submitted = true;
+    if (this.user.password != this.confirmpassword) {
+      alert("Passwords don't match!");
+    }
+    else if (this.form2.invalid) {
+      return;
+    }
+    else {
+      //this.user.status = "PENDING";
+      this.user.user_type = "BIDDER";
+      localStorage.setItem('usertype', this.user.user_type);
+
+      console.log(this.user);
+
+      this.service.signupfarmer(this.user).subscribe(data => {
+
         //this.message=data.message;
         alert(JSON.stringify(data));
-        if(data.status=="SUCCESS"){
-          localStorage.setItem('user_id',data.registeredCustomerId);
+        if (data.status == "SUCCESS") {
+          localStorage.setItem('user_id', data.registeredCustomerId);
           this.router.navigate(['/signup-documents']);
         }
-  })
-}
+      })
+
+      this.submitted = true;
+      if (this.form2.invalid) {
+        return;
+      }
+    }
+
+
+  }
 
 }
